% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to a dimple chart}
\usage{
add_legend(viz, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{width}{}

\item{height}{}

\item{horizontalAlign}{}
}
\description{
Parameters provided should match those for \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.legend}{dimple.legend constructor}.
  If incomplete parameters are provided then some sane defaults will be applied.
}
\examples{
dimple( weight ~ Time, groups = "Diet" , type = "line", data = ChickWeight, width = 600, height = 400 ) \%>\%
  add_legend( )

dimple( weight ~ Time, groups = "Diet" , type = "line", data = ChickWeight, width = 600, height = 400 ) \%>\%
  add_legend( x = 50, y = 390, width = 500, horizontalAlign = "right" )
}
\seealso{
Other dimple: \code{\link{add_title}};
  \code{\link{default_colors}}; \code{\link{dimple}};
  \code{\link{facet.dimple}}; \code{\link{pAxis.dimple}},
  \code{\link{xAxis.dimple}}, \code{\link{yAxis.dimple}},
  \code{\link{zAxis.dimple}}; \code{\link{set_bounds}}
}

