% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{add_series}
\alias{add_series}
\title{Add a series to your dimple chart}
\usage{
add_series(viz, ...)
}
\arguments{
\item{viz}{\code{rCharts2} chart to which you would like to add a series}
}
\description{
Add a series to your dimple chart
}

