% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{add_title}
\alias{add_title}
\title{Add a title to your chart}
\usage{
add_title(viz, text = NULL, html = NULL, ...)
}
\arguments{
\item{viz}{\code{rCharts2} chart to which you would like to add a title}

\item{text}{\code{string} of text to provide in the title}

\item{html}{\code{string} of html for your title.  Providing \code{html} instead of
\code{text} to the title allows the potential to custom format your
title or provide multiple lines.}

\item{...}{other arguments for consistency with \code{rCharts2} methods.}
}
\description{
Add a title to your chart
}
\seealso{
Other dimple: \code{\link{add_legend}};
  \code{\link{default_colors}}; \code{\link{dimple}};
  \code{\link{facet.dimple}}; \code{\link{pAxis.dimple}},
  \code{\link{xAxis.dimple}}, \code{\link{yAxis.dimple}},
  \code{\link{zAxis.dimple}}; \code{\link{set_bounds}}
}

