% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{default_colors}
\alias{default_colors}
\title{Modify colors used in Dimple charts}
\usage{
default_colors(viz, ...)
}
\arguments{
\item{viz}{a dimple object}

\item{colors}{a vector of colors in hex or a \code{\link[htmlwidgets]{JS}} function, such as a \href{https://github.com/mbostock/d3/wiki/Ordinal-Scales#categorical-colors}{d3 color scale}.}
}
\description{
Modify colors used in Dimple charts
}
\examples{
ChickWeight \%>\%
  dimple( x = c("Time", "Diet"), y = "weight", groups = "Diet", type = "bar" ) \%>\%
  defaultColors( rainbow(10) )

ChickWeight \%>\%
  dimple( x = c("Time", "Diet"), y = "weight", groups = "Diet", type = "bar" ) \%>\%
  defaultColors( htmlwidgets::JS( 'd3.scale.category10()' ) )
}
\seealso{
Other dimple: \code{\link{add_legend}};
  \code{\link{add_title}}; \code{\link{colorAxis}};
  \code{\link{dimple}}; \code{\link{facet.dimple}};
  \code{\link{set_bounds}}; \code{\link{xAxis.dimple}},
  \code{\link{yAxis.dimple}}, \code{\link{zAxis.dimple}}
}

