% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{pAxis}
\alias{pAxis}
\title{Modify plot elements that relate to the p-axis.}
\usage{
pAxis(x, ...)
}
\description{
This is an S3 method.
}
\seealso{
Other axis: \code{\link{pAxis.dimple}},
  \code{\link{xAxis.dimple}}, \code{\link{yAxis.dimple}},
  \code{\link{zAxis.dimple}}; \code{\link{yAxis}};
  \code{\link{zAxis}}
}

