% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{set_bounds}
\alias{set_bounds}
\title{Set the \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.chart#setBounds}{bounds} or size of the chart within the svg container.}
\usage{
set_bounds(viz, x, y, width, height)
}
\arguments{
\item{viz}{a dimple object}

\item{x}{value in pixels where the chart starts on the left}

\item{y}{value in pixels where the chart starts on the bottom}

\item{width}{value in pixels of the width of chart.  Note this is not
the right edge.  Rather the right edge will be x + width.}

\item{height}{height value in pixels of the height of chart.  Note this is
 not the top.  Rather the top will be y + height.}
}
\description{
The chart size can be set independently from the size of the containing node.
}
\examples{
ChickWeight \%>\%
  dimple( weight ~ Time, groups = "Diet", type = "bubble", width = 600, height = 600 ) \%>\%
  setBounds( x = 200, y = 100, width = 250, height = 300 )
}
\seealso{
Other dimple: \code{\link{add_legend}};
  \code{\link{add_title}}; \code{\link{default_colors}};
  \code{\link{dimple}}; \code{\link{facet.dimple}};
  \code{\link{pAxis.dimple}}, \code{\link{xAxis.dimple}},
  \code{\link{yAxis.dimple}}, \code{\link{zAxis.dimple}}
}

