% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{xAxis.dimple}
\alias{xAxis.dimple}
\alias{yAxis.dimple}
\alias{zAxis.dimple}
\title{Modify plot elements that relate to the x and y axes of a dimple chart}
\usage{
\method{xAxis}{dimple}(viz, ...)

\method{yAxis}{dimple}(viz, ...)

\method{zAxis}{dimple}(viz, ...)
}
\arguments{
\item{...}{other properties for dimple axis.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#properties-1}{Properties}.}

\item{type}{"addMeasureAxis", "addCategoryAxis", "addTimeAxis", or "addPctAxis"}

\item{orderRule}{defines how to sort a categorical axis.  \code{orderRule} can either be a field name as a string or a vector of values in the desired order.}

\item{grouporderRule}{defines how to the values in a group.  \code{orderRule} can either be a field name as a string or a vector of values in the desired order.}

\item{overrideMin}{manually specify a minimum axis value.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#overrideMin}{dimple.axis.overrrideMin}.}

\item{overrideMax}{manually specify a maximum axis value.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#overrideMax}{dimple.axis.overrrideMax}.}

\item{inputFormat}{the d3 time format of the values in a field if an axis is \code{addTimeAxis}.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#dateParseFormat}{dimple.axis.dateParseFormat}.}

\item{outputFormat}{the d3 format of an axis.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#tickFormat}{dimple.axis.tickFormat}.}

\item{timePeriod}{for time axis periodicity of ticks. See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#timePeriod}{dimple.axis.timePeriod}.}

\item{timeInterval}{for time axis number of periods between ticks.  See \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis#timeInterval}{dimple.axis.timeInterval}.}

\item{fontSize}{for font size of axis.  Any valid CSS \code{font-size} will work.}

\item{fontFamily}{for font size of axis.  Any valid CSS \code{font-family} will work.}

\item{useLog}{for logarithmic axis.}

\item{logBase}{for base of log if axis is logarithmic \code{useLog = TRUE}.}
}
\description{
\code{xAxis}, \code{yAxis}, \code{zAxis} configure the x, y, and z axis properties of the chart.
In most cases, parameters closely match those of \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.axis}{dimple.axis}.
}
\seealso{
Other axis: \code{\link{yAxis}}; \code{\link{zAxis}}

Other dimple: \code{\link{add_legend}};
  \code{\link{add_title}}; \code{\link{default_colors}};
  \code{\link{dimple}}; \code{\link{facet.dimple}};
  \code{\link{set_bounds}}
}

