% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{xAxis}
\alias{xAxis}
\title{Modify plot elements that relate to the x-axis.}
\usage{
xAxis(x, ...)
}
\description{
This is an S3 method.
}
\seealso{
Other xAxis: \code{\link{facet}};
  \code{\link{xAxis.dimple}}, \code{\link{yAxis.dimple}},
  \code{\link{zAxis.dimple}}; \code{\link{yAxis}};
  \code{\link{zAxis}}
}

