% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimple.R
\name{dimple}
\alias{dimple}
\title{Plot R data with dimple}
\usage{
dimple(data, x, type = type, y = NULL, z = NULL, groups = NULL,
  group = NULL, storyboard = NULL, ..., pipe = F, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} with the data to plot}

\item{x}{formula \code{y ~ x} or column name(s) of \code{x} variable.  Dimple can accept either a single or multiple column name(s)
as \code{x} or \code{y}.  If multiple column names for either \code{x} or \code{y}, then the formula method cannot be used.}

\item{type}{a string indicating plot type ( \href{http://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.plot}{bar, line, area, or bubble} )}

\item{y}{column name(s) of \code{y} variable if \code{x} is not a \code{formula}.  Dimple can accept either a single or multiple column name(s)
as \code{y} or \code{y}.  If multiple column names for either \code{x} or \code{y}, then the formula method cannot be used.}

\item{groups}{see \code{group}}

\item{group}{column name(s) of \code{groups}.  Dimple differs from other \code{rCharts2} libraries
in its ability to handle multiple fields or column names as \code{groups}.}

\item{storyboard}{column name(s) with data for \href{https://github.com/PMSI-AlignAlytics/dimple/wiki/dimple.storyboard}{dimple.storyboard}}

\item{...}{various additional arguments to future-proof and provide
undocumented functionality.}

\item{pipe}{\code{logical} to return a \code{\link[pipeR]{Pipe}}.}

\item{width}{width in pixels of the plot container.  Plot \code{width} can be set separately (see \code{\link{setBounds}}).}

\item{height}{height in pixels of the plot container.  Plot \code{height} can be set separately (see \code{\link{setBounds}}).}

\item{elementId}{\code{string} css selector id for the chart.  If the \code{elementId} is \code{NULL},
          a semi-random id will be assigned.}
}
\description{
Use \href{http://dimplejs.org}{Dimple} with this \code{htmlwidget}.  Dimple is a powerful javascript library
 based on \href{http://d3js.org}{d3.js}.  Dimple supports numerous chart \href{http://dimplejs.org/examples_index.html}{types} and is
 well \href{https://github.com/PMSI-AlignAlytics/dimple/wiki}{documented}.
}
\seealso{
Other dimple: \code{\link{add_legend}};
  \code{\link{default_colors}}; \code{\link{facet.dimple}};
  \code{\link{set_bounds}}; \code{\link{xAxis.dimple}},
  \code{\link{yAxis.dimple}}, \code{\link{zAxis.dimple}}
}

